ui restructure

+++ backend/app/routers/utils.py
@@ -1,14 +1,16 @@
 from fastapi import FastAPI, File, UploadFile, HTTPException, APIRouter
 from fastapi.responses import JSONResponse
 import shutil, logging, os
-from typing import List
+from typing import List, Optional
 
 logging.basicConfig(level=logging.INFO)
 
 utils_router = APIRouter(tags=["Utilities"])
 
 @utils_router.post("/upload/")
-async def upload_files(train_file: UploadFile = File(...), test_file: UploadFile = File(...)):
+async def upload_files(train_file: UploadFile = File(...), 
+                       test_file: UploadFile = File(...), 
+                       grammar: UploadFile = File(...)):
    try:
        train_dir = "../examples"
        if not os.path.exists(train_dir):
@@ -16,6 +18,7 @@ async def upload_files(train_file: UploadFile = File(...), test_file: UploadFile
        
        train_path = os.path.join(train_dir, train_file.filename)
        test_path = os.path.join(train_dir, test_file.filename)
+       grammar_path = os.path.join(train_dir, grammar.filename)
 
        # Check if the train file already exists
        if os.path.exists(train_path):
@@ -25,6 +28,9 @@ async def upload_files(train_file: UploadFile = File(...), test_file: UploadFile
        if os.path.exists(test_path):
            raise HTTPException(status_code=400, detail=f"Test file '{test_file.filename}' already exists.")
        
+       if os.path.exists(grammar_path):
+           raise HTTPException(status_code=400, detail=f"Grammar file '{grammar.filename}' already exists.")
+       
        # Save train file
        with open(train_path, "wb") as buffer:
            shutil.copyfileobj(train_file.file, buffer)
@@ -32,6 +38,9 @@ async def upload_files(train_file: UploadFile = File(...), test_file: UploadFile
        # Save test file
        with open(test_path, "wb") as buffer:
            shutil.copyfileobj(test_file.file, buffer)
+         
+       with open(grammar_path, "wb") as buffer:
+           shutil.copyfileobj(grammar.file, buffer)
 
        logging.info(f"Files {train_file.filename} and {test_file.filename} uploaded successfully.")
        return {"message": "Files uploaded successfully"}
@@ -42,17 +51,16 @@ async def upload_files(train_file: UploadFile = File(...), test_file: UploadFile
        logging.error(f"Error uploading files: {str(e)}")
        raise HTTPException(status_code=500, detail=f"Error uploading files: {str(e)}")
    
-
-   ###### Make an endpoint to get all the files in the examples folder and then enother one to delete a file that is no longer needed
-   
+    
+    
 @utils_router.get("/files/", response_model=List[str])
 async def list_files():
     try:
-        file_dir = "../examples"
+        file_dir =  "../examples"
         if not os.path.exists(file_dir):
             return []
         
-        files = [f for f in os.listdir(file_dir) if f.endswith(('.train', '.test'))]
+        files = [f for f in os.listdir(file_dir) if f.endswith(('.train', '.test', '.bnf'))]
         return files
     except Exception as e:
         logging.error(f"Error listing files: {str(e)}")
